+++ server/src/main/java/org/opensearch/index/engine/MergedSegmentWarmer.java
@@ -8,11 +8,14 @@
 
 package org.opensearch.index.engine;
 
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.message.ParameterizedMessage;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.LeafReader;
 import org.apache.lucene.index.SegmentCommitInfo;
 import org.apache.lucene.index.SegmentReader;
 import org.opensearch.cluster.service.ClusterService;
+import org.opensearch.common.logging.Loggers;
 import org.opensearch.index.shard.IndexShard;
 import org.opensearch.indices.recovery.RecoverySettings;
 import org.opensearch.transport.TransportService;
@@ -20,17 +23,20 @@ import org.opensearch.transport.TransportService;
 import java.io.IOException;
 
 /**
- * Implementation of a {@link IndexWriter.IndexReaderWarmer} when local on-disk segment replication is enabled.
+ * Implementation of a {@link IndexWriter.IndexReaderWarmer} for merged segment replication in
+ * local on-disk and remote store enabled domains.
  *
  * @opensearch.internal
  */
-public class LocalMergedSegmentWarmer implements IndexWriter.IndexReaderWarmer {
+public class MergedSegmentWarmer implements IndexWriter.IndexReaderWarmer {
     private final TransportService transportService;
     private final RecoverySettings recoverySettings;
     private final ClusterService clusterService;
     private final IndexShard indexShard;
 
-    public LocalMergedSegmentWarmer(
+    private final Logger logger;
+
+    public MergedSegmentWarmer(
         TransportService transportService,
         RecoverySettings recoverySettings,
         ClusterService clusterService,
@@ -40,14 +46,30 @@ public class LocalMergedSegmentWarmer implements IndexWriter.IndexReaderWarmer {
         this.recoverySettings = recoverySettings;
         this.clusterService = clusterService;
         this.indexShard = indexShard;
+        this.logger = Loggers.getLogger(getClass(), indexShard.shardId());
     }
 
     @Override
     public void warm(LeafReader leafReader) throws IOException {
         // IndexWriter.IndexReaderWarmer#warm is called by IndexWriter#mergeMiddle. The type of leafReader should be SegmentReader.
         assert leafReader instanceof SegmentReader;
+        assert indexShard.indexSettings().isSegRepLocalEnabled() || indexShard.indexSettings().isRemoteStoreEnabled();
 
+        long startTime = System.currentTimeMillis();
         SegmentCommitInfo segmentCommitInfo = ((SegmentReader) leafReader).getSegmentInfo();
+        logger.trace(() -> new ParameterizedMessage("Warming segment: {}", segmentCommitInfo));
         indexShard.publishMergedSegment(segmentCommitInfo);
+        logger.trace(() -> {
+            long segmentSize = -1;
+            try {
+                segmentSize = segmentCommitInfo.sizeInBytes();
+            } catch (IOException ignored) {}
+            return new ParameterizedMessage(
+                "Completed segment warming for {}. Size: {}B, Timing: {}ms",
+                segmentCommitInfo.info.name,
+                segmentSize,
+                (System.currentTimeMillis() - startTime)
+            );
+        });
     }
 }